requirements.txt
langchain==0.0.198
openai==0.27.7
faiss-cpu==1.7.4
tiktoken==0.4.0
pdf2image==1.16.3
unstructured==0.7.6
python-dotenv==1.0.0